appendCard 네이밍 변경, CardBox 클래스 추가, Access Control 재설정, README.md 작성
- getCard -> appendCard
- CardBox 클래스 추가
- Access Control 재 설정(Player.cards) - private는 특정 엔티티가 선언된 괄호 안에서만 사용 가능(엔티티: 데이터의 집합)
